requirements.txt
pymongo[srv]
sentence-transformers
fastapi
uvicorn
python-dotenv
streamlit
requests
pandas